backend/requirements.txt
fastapi
uvicorn[standard]
pymysql
python-dotenv

